[APFM-5404] Fix annotation line number (#495)
There was a bug in the annotations reported. Previously if the coverage was missing for ranges e.g 25-30, 35-50, the annotation reported mentioned coverage failed between lines 25-50 which is misleading. The current fix reports annotations coverage failed between lines 25-30 and coverage failed between lines 35-50.

The issue popped up when we exclude certain methods from code coverage, their lines won't be reported in the XML file and hence the bug was not discovered previously.